requirements.txt
django-nested-admin
django-grappelli
django-bower
